Latest appng

+++ appng-appngizer/src/main/java/org/appng/appngizer/controller/SiteController.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2017 the original author or authors.
+ * Copyright 2011-2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,14 +15,25 @@
  */
 package org.appng.appngizer.controller;
 
+import java.io.File;
+import java.io.IOException;
+import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+import org.apache.commons.io.FileUtils;
 import org.appng.api.BusinessException;
+import org.appng.api.Environment;
+import org.appng.api.RequestUtil;
+import org.appng.api.SiteProperties;
 import org.appng.api.messaging.Messaging;
 import org.appng.api.messaging.Sender;
+import org.appng.api.model.Site.SiteState;
 import org.appng.api.support.FieldProcessorImpl;
 import org.appng.api.support.environment.DefaultEnvironment;
+import org.appng.appngizer.model.Link;
 import org.appng.appngizer.model.Site;
 import org.appng.appngizer.model.Sites;
 import org.appng.core.controller.messaging.ReloadSiteEvent;
@@ -31,21 +42,25 @@ import org.slf4j.Logger;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import lombok.extern.slf4j.Slf4j;
 
+@Controller
 @Slf4j
 @RestController
 public class SiteController extends ControllerBase {
 
-	@RequestMapping(value = "/site", method = RequestMethod.GET)
+	@GetMapping(value = "/site")
 	public ResponseEntity<Sites> listSites() {
-		List<Site> siteList = new ArrayList<Site>();
+		List<Site> siteList = new ArrayList<>();
 		for (SiteImpl site : getCoreService().getSites()) {
 			Site fromDomain = Site.fromDomain(site);
 			fromDomain.applyUriComponents(getUriBuilder());
@@ -56,38 +71,79 @@ public class SiteController extends ControllerBase {
 		return new ResponseEntity<Sites>(sites, HttpStatus.OK);
 	}
 
-	@RequestMapping(value = "/site/{name}", method = RequestMethod.GET)
+	@GetMapping(value = "/site/{name}")
 	public ResponseEntity<Site> getSite(@PathVariable("name") String name) {
 		SiteImpl site = getSiteByName(name);
 		if (null == site) {
 			return notFound();
 		}
 		Site fromDomain = Site.fromDomain(site);
+		if (null != getSender(DefaultEnvironment.getGlobal()) || supportsReloadFile(site)) {
+			fromDomain.addLink(new Link("reload", "/site/" + name + "/reload"));
+		}
 		fromDomain.applyUriComponents(getUriBuilder());
 		return ok(fromDomain);
 	}
 
-	@RequestMapping(value = "/site/{name}/reload", method = RequestMethod.PUT)
-	public ResponseEntity<Void> reloadSite(@PathVariable("name") String name) {
-		Sender sender = Messaging.getMessageSender(DefaultEnvironment.get(context));
-		if (null == sender) {
+	public static class ReloadSiteFromAppNGizer extends ReloadSiteEvent implements Serializable {
+
+		public ReloadSiteFromAppNGizer(String siteName) {
+			super(siteName);
+		}
+
+		@Override
+		protected void setNodeId(String nodeId) {
+			super.setNodeId(nodeId + "_appNGizer");
+		}
+
+	}
+
+	@PutMapping(value = "/site/{name}/reload")
+	public ResponseEntity<Void> reloadSite(@PathVariable("name") String name) throws BusinessException {
+		SiteImpl site = getSiteByName(name);
+		if (null == site) {
 			return notFound();
 		}
-		sender.send(new ReloadSiteEvent(name));
+		Sender sender = getSender(DefaultEnvironment.getGlobal());
+		if (null != sender) {
+			LOGGER.debug("messaging is active, sending ReloadSiteEvent");
+			sender.send(new ReloadSiteFromAppNGizer(name));
+		} else if (supportsReloadFile(site)) {
+			String rootDir = site.getProperties().getString(SiteProperties.SITE_ROOT_DIR);
+			File reloadFile = new File(rootDir, ".reload");
+			try {
+				LOGGER.debug("Created reload marker {}", reloadFile.getAbsolutePath());
+				FileUtils.touch(reloadFile);
+			} catch (IOException e) {
+				throw new BusinessException(e);
+			}
+		} else {
+			return reply(HttpStatus.METHOD_NOT_ALLOWED);
+		}
 		return ok(null);
 	}
 
-	@RequestMapping(value = "/site", method = RequestMethod.POST)
-	public ResponseEntity<Site> createSite(@RequestBody org.appng.appngizer.model.xml.Site site) {
-		SiteImpl currentSite = getSiteByName(site.getName());
-		if (null != currentSite) {
-			return conflict();
+	private Sender getSender(Environment env) {
+		return Messaging.getMessageSender(env);
+	}
+
+	private boolean supportsReloadFile(org.appng.api.model.Site site) {
+		return Boolean.TRUE.equals(site.getProperties().getBoolean(SiteProperties.SUPPORT_RELOAD_FILE));
+	}
+
+	@PostMapping(value = "/site")
+	public ResponseEntity<Site> createSite(@RequestBody org.appng.appngizer.model.xml.Site site,
+			HttpServletRequest request) throws ConflictException {
+		SiteImpl siteToCreate = Site.toDomain(site);
+		List<String> conflictMsgs = new ArrayList<>();
+		if (getCoreService().checkSiteNameConflicts(siteToCreate, "all", request.getLocale(), conflictMsgs)) {
+			throw new ConflictException(conflictMsgs);
 		}
 		getCoreService().createSite(Site.toDomain(site));
 		return created(getSite(site.getName()).getBody());
 	}
 
-	@RequestMapping(value = "/site/{name}", method = RequestMethod.PUT)
+	@PutMapping(value = "/site/{name}")
 	public ResponseEntity<Site> updateSite(@PathVariable("name") String name,
 			@RequestBody org.appng.appngizer.model.xml.Site site) {
 		SiteImpl siteByName = getSiteByName(name);
@@ -96,17 +152,30 @@ public class SiteController extends ControllerBase {
 		}
 		siteByName.setHost(site.getHost());
 		siteByName.setDomain(site.getDomain());
+		siteByName.getHostAliases().clear();
+		if (null != site.getHostAliases()) {
+			if(null == siteByName.getHostAliases()) {
+				siteByName.setHostAliases(new HashSet<>());
+			}
+			siteByName.getHostAliases().addAll(site.getHostAliases().getAlias());
+		}
 		siteByName.setDescription(site.getDescription());
 		siteByName.setActive(site.isActive());
 		getCoreService().saveSite(siteByName);
 		return ok(getSite(name).getBody());
 	}
 
-	@RequestMapping(value = "/site/{name}", method = RequestMethod.DELETE)
+	@DeleteMapping(value = "/site/{name}")
 	public ResponseEntity<Void> deleteSite(@PathVariable("name") String name) throws BusinessException {
 		SiteImpl currentSite = getSiteByName(name);
 		if (null == currentSite) {
 			return notFound();
+		}
+		Environment environment = DefaultEnvironment.getGlobal();
+		org.appng.api.model.Site site = RequestUtil.getSiteByName(environment, name);
+		if (site != null && (site.getState() == SiteState.STARTING || site.getState() == SiteState.STARTED
+				|| site.getState() == SiteState.SUSPENDED || site.getState() == SiteState.STOPPING)) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
 		}
 		getCoreService().deleteSite(name, new FieldProcessorImpl("delete-site"));
 		HttpHeaders headers = new HttpHeaders();
@@ -115,7 +184,7 @@ public class SiteController extends ControllerBase {
 	}
 
 	Logger logger() {
-		return log;
+		return LOGGER;
 	}
 
 }
